Revert "Db_customer modified and tested, address partly completed"
This reverts commit c38ccae96e859eb2223ddcabf07fdd7688f37975.

+++ myshop-domain/src/main/java/com/db/entities/Db_Customer.java
@@ -28,7 +28,7 @@ public class Db_Customer {
 	@Column(name="password")
 	private String password;
 
-	public int getRid() {
+	public long getRid() {
 		return rid;
 	}
